Add verbose option to do_cmd
This commit introduces the option to toggle do_cmd verobsity output of
the bash commands that it executes on the target server.

This was implemented because you don't always need or want to see the
bash command that was executed by do_cmd. The output is only really
useful for development and or troubleshooting.

Errors from do_cmd were not silenced because they shoud print by
default. You need to know if the bash command fails.