Refactor - Re-implemented vhost directive crawler methods
Implemented the _find_directive method which checks if a requested
vhost directive is set in the parsed vhost config. Each directive
crawler method calls the _find_directive method to validate if the
vhost directive is set and if not returns 'not set' to the directive
crawler methods.

Created the following directive crawler methods:

Top-Down Order:
-ServerName
-ServerAlias
-DocumentRoot
-ErrorLog
-CustomLog