Enhance - vhost conf path affiliation with vhost_objects
Reimplemented main() method for loop that iterates over vhost objects. Modified
the method so that it loops over a list of configuration paths instead of vhost
objects.

Removed the internal for loop in vhosts_extraction method. This changes
the behavior of the vhosts_extraction method in that the method now only takes one configuration
argument at a time, instead of a list of arguments. This modification
along with the re-implementation of the for loop ensures that the
configuration paths and vhost objects match correctly.