Enhance - Updated  methods & test cases to return None
Per the PEP8 guidelines, best practice is to return None when a
function returns nothing. Following the guideline and Evan's
recommendation, I have implemented the changes to the try/catch block in
the _find_directive method in domain_recon.

Additionally, the unitTest cases in test_domain_recon have been changed
to check for None instead of 'Not Set'.